requirements.txt
langchain==0.0.178
pandas==2.0.1
Pillow==9.5.0
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.22.0
